WIP of nicer parser
Trrying to work out some nice abstractions for parsing this info. Not
sure I'm quite there with this. I basically want to be able to define
structs for records with only the information I care about, but not need
to write custom parsing for each one. Here I'm writing a mapping from
field num to pointer to field for each one I care about but I think I
could autogenerate this somehow.