Improve the comptime parsing stuff
Effectively move the mappings struct inline into an assignField
function. This function gets all the fields from the destination struct
that match the required type. It then gets the name of the field id from
defs.zig. It then loops through the fields with the right type and
checks if its name matches the name we want.